minor changes and deletions

+++ mathFunctions/furtherMathFunctions.h
@@ -141,7 +141,7 @@ enum estimatedFunction {FUNCTION_CODE_SPHERICAL, FUNCTION_CODE_LINEAR, FUNCTION_
                                                                   int maxIterationsNr, double myEpsilon,
                                                                   std::vector <double>& x, std::vector<double>& y,
                                                                   std::vector<double>& weights);
-        int bestFittingMarquardt_nDimension_singleFunction(double (*func)(double, std::vector<double>&),
+        double bestFittingMarquardt_nDimension_singleFunction(double (*func)(double, std::vector<double>&),
                                                            int nrTrials, int nrMinima,
                                                            std::vector <double>& parametersMin, std::vector <double>& parametersMax,
                                                            std::vector <double>& parameters, std::vector <double>& parametersDelta,
